refactor: 파일 이름 변경, [상품 추가] 이벤트 리팩토링
- validate 파일 생성 및 모듈화
- 이벤트 코드 모듈화
