+++ src/js/controller/productAddEvent.mjs
@@ -1,26 +1,27 @@
 import { renderProductAddPage, renderProductStatusTable } from '../view/index.mjs';
 import { Product } from '../model/Product.mjs';
-import { products, addProduct } from '../model/products.mjs';
+import { addProduct } from '../model/setProducts.mjs';
+import { inputValidate } from './util/validate.mjs';
 
-function renderProductAddTab() {
+function initProductAddPage() {
   document.querySelector('main').remove();
-  renderProductAddPage(products);
+  renderProductAddPage();
+}
+
+function productAddButton(e) {
+  const name = document.querySelector('#product-name-input').value;
+  const price = document.querySelector('#product-price-input').value;
+  const quantity = document.querySelector('#product-quantity-input').value;
+
+  if (e.target !== document.querySelector('#product-add-button')) return;
+  if (!inputValidate(name, price, quantity)) return;
+
+  addProduct(new Product(name, price, quantity));
+  renderProductStatusTable();
 }
 
 export function productAddEvent() {
   const $productAddMenu = document.querySelector('#product-add-menu');
-
-  $productAddMenu.addEventListener('click', renderProductAddTab);
-
-  window.addEventListener('click', e => {
-    if (e.target !== document.querySelector('#product-add-button')) return;
-
-    e.preventDefault();
-    const name = document.querySelector('#product-name-input').value;
-    const price = document.querySelector('#product-price-input').value;
-    const quantity = document.querySelector('#product-quantity-input').value;
-
-    addProduct(new Product(name, price, quantity));
-    renderProductStatusTable();
-  });
+  $productAddMenu.addEventListener('click', initProductAddPage);
+  window.addEventListener('click', e => productAddButton(e));
 }
