+++ src/js/model/setProducts.mjs
